continued fixing dh

+++ diffieHelmanPage/diffieHelmanPageCommander.cpp
@@ -42,7 +42,7 @@ void DiffieHellmanPageCommander::createPage() {
         auto* self = static_cast<DiffieHellmanPageCommander*>(t->user_data);
         self->poll();
         
-        if (self->currentSoldierIndex >= self->allSoldiersVector.size())
+        if (self->currentSoldierIndex >= (int)self->allSoldiersVector.size())
         {
             Serial.printf("%d %d\n", self->currentSoldierIndex, self->allSoldiersVector.size());
             Serial.println("Releasing timer!");
@@ -112,6 +112,7 @@ void DiffieHellmanPageCommander::startProcess() {
 
 void DiffieHellmanPageCommander::startExchangeWithNextSoldier() {
     Serial.println("startExchangeWithNextSoldier");
+
     if (currentSoldierIndex >= allSoldiersVector.size()) {
         setStatusText("✅ All exchanges complete");
         commanderProcessStarted = false;
